I wanted to get rid of the sub minute field as that may cause drift between accumulated deltas and the absolute value. Then I realized that sub minutes aren't used anywhere. So I moved to one double-typed value for minutes. All time handling has been moved to use doubles now. Savegames are invalid because of type change. It made me realize that I should move to Double everywhere ... especially on coordinates.

+++ Game/Data/Data.cs
@@ -65,10 +65,10 @@ namespace ButtonOffice
         public static readonly Single LampThreeX = 11.0f;
         public static readonly Single LampTwoX = 6.0f;
         public static readonly Single LampWidth = 3.0f;
-        public static readonly Single MeanMinutesToBrokenComputer = 500.0f;
-        public static readonly Single MeanMinutesToBrokenLamp = 1500.0f;
+        public static readonly Double MeanMinutesToBrokenComputer = 500.0f;
+        public static readonly Double MeanMinutesToBrokenLamp = 1500.0f;
         public static readonly UInt64 NewGameCents = 5000000ul;
-        public static readonly UInt64 NewGameMinutes = 480ul;
+        public static readonly Double NewGameMinutes = 480.0;
         public static readonly Int32 NewGameWorldBlockHeight = 136;
         public static readonly Int32 NewGameWorldBlockWidth = 132;
         public static readonly Color OfficeBackgroundColor = Color.FromArgb(255, 255, 255, 255);
